(1.20.4) Remove custom name from chairs
Okay, so the problem is that in 1.20.4 Mojang decided that every non-mob entity should show their nametag. WHY?!?!??!

There isn't a gamerule or tag to disable this or anything.

I had to get rid of everything that names the interaction entities. Hopefully this does not affect anything else.

However, existing entities will still be named like this. I have to make a command to remove the name out of everything (which is apparently impossible? https://bugs.mojang.com/browse/MC-267259)